database models refactor
